start splitting out page-specific css
http/2 means we don't need to bundle all the styles together. Splitting 
out large page-specific styles means we'll be able to flatten them and 
decrease their size. Still some duplication to get rid of (e.g. 
typography) but this is a start. Probably also possible to write a 
function to handle the brunch css config automatically instead of 
manually.