Rework header
* Trying to merge the masthead and the hero header into the same grid 
was a very smart thing to do (as in "ingenious") and also a very stupid 
thing to do (as in "not wise").
* Putting the masthead over the hero image makes it illegible anyway 
unless we do some not-so-pretty stuff. The masthead shouldn't scream at 
you.
* Separating out the masthead and the header means we can have golden 
grids at a lower level in the page structure, which means less 
pulling-up-things to place them on the grid.
* Also fix some sub-titles and highlights