copy rest compiler plugin to start adapting it
To output srcset elements in our HTML, we need to override the HTML 
writer and change the output of the figure directive. Right now this 
means that the built-in reST plugins (like doc, post_list etc) don't 
work anymore with this custom compiler, but that's something we can fix. 
At least, this compiles.