f
* srcset is an img attribute, so it's actually visit_image that we want 
to override.
* visit_image in html5_polyglot only does video stuff and delegates 
everything else to the HTMLWriter in its parent class _html_base. What 
we do here, since our translator inherits from the html5_polyglot 
HTMLTranslator, is that we override visit_image, and instead of calling 
its super for non-video stuff, we call a copy of it located right here 
in our translator class.
* Copying those entire functions isn't ideal, so to make this as 
future-proof as possible, we change very little; the main goal is to be 
able to add attributes to the img tag.
* For now, we're not actually adding any attributes, but this first step 
reproduces existing functionality while making customization possible